refactor: split browser oauth code from shared atreply client
